consolidating and reorganizing the files, moving embed to embedding
