--- examples/cli/env.example
@@ -1,18 +1,0 @@
-# API Keys for various AI providers
-# Copy this file to .env and add your actual API keys
-
-# OpenAI API Key
-# Get yours at: https://platform.openai.com/api-keys
-OPENAI_API_KEY=sk-...
-
-# Anthropic API Key  
-# Get yours at: https://console.anthropic.com/settings/keys
-ANTHROPIC_API_KEY=sk-ant-...
-
-# Google Gemini API Key
-# Get yours at: https://aistudio.google.com/app/apikey
-GOOGLE_API_KEY=...
-
-# Ollama Configuration (for local models)
-# Default: http://localhost:11434
-# OLLAMA_HOST=http://localhost:11434
