--- examples/cli/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "@tanstack/ai-cli-demo",
-  "version": "0.1.0",
-  "private": true,
-  "description": "CLI demo for TanStack AI library",
-  "type": "module",
-  "module": "dist/index.js",
-  "bin": {
-    "tanstack-ai": "./dist/index.js"
-  },
-  "scripts": {
-    "build": "tsdown",
-    "dev:cli": "tsx src/index.ts",
-    "dev:watch": "tsx watch src/index.ts",
-    "start": "node dist/index.js",
-    "clean": "rm -rf dist node_modules",
-    "chat": "tsx src/index.ts chat",
-    "generate": "tsx src/index.ts generate",
-    "summarize": "tsx src/index.ts summarize",
-    "embed": "tsx src/index.ts embed",
-    "tools": "tsx src/index.ts tools"
-  },
-  "keywords": [
-    "ai",
-    "cli",
-    "demo",
-    "tanstack"
-  ],
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "@tanstack/ai": "workspace:*",
-    "@tanstack/ai-anthropic": "workspace:*",
-    "@tanstack/ai-gemini": "workspace:*",
-    "@tanstack/ai-ollama": "workspace:*",
-    "@tanstack/ai-openai": "workspace:*",
-    "chalk": "^5.3.0",
-    "commander": "^12.0.0",
-    "dotenv": "^16.4.5",
-    "inquirer": "^9.2.19",
-    "ora": "^8.0.1"
-  },
-  "devDependencies": {
-    "@types/inquirer": "^9.0.7",
-    "@types/node": "^22.10.2",
-    "tsdown": "^0.15.9",
-    "tsx": "^4.7.1",
-    "typescript": "^5.7.2"
-  }
-}
