--- examples/cli/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src",
-    "types": ["node"],
-  },
-  "include": ["src/**/*.ts", "src/**/*.tsx"],
-  "exclude": ["node_modules", "dist", "**/*.config.ts"],
-  "references": [
-    { "path": "../../packages/typescript/ai" },
-    { "path": "../../packages/typescript/ai-openai" },
-    { "path": "../../packages/typescript/ai-anthropic" },
-    { "path": "../../packages/typescript/ai-ollama" },
-    { "path": "../../packages/typescript/ai-gemini" }
-  ]
-}
