+++ packages/typescript/ai/src/index.ts
@@ -1,4 +1,6 @@
-export { chat, summarize, embed } from "./standalone-functions";
+export { chat } from "./chat";
+export { summarize } from "./summarize";
+export { embedding } from "./embedding";
 export { tool } from "./tool-utils";
 export {
   responseFormat,
@@ -10,7 +12,7 @@ export {
   toStreamResponse,
 } from "./stream-to-response";
 export { BaseAdapter } from "./base-adapter";
-export { ToolCallManager } from "./tool-call-manager";
+export { ToolCallManager } from "./tool-calls";
 export {
   maxIterations,
   untilFinishReason,
