+++ packages/typescript/tests-adapters/README.md
@@ -13,7 +13,7 @@ This package contains simple tests for the four AI adapter libraries:
 - Tools: Uses the `get_temperature` tool and checks for "70"/"seventy" and tool wiring.
 - Approval: Requires approval for `addToCart` and ensures the tool executes once.
 - Summarize: Summarizes a short paragraph about Paris.
-- Embed: Generates embeddings for two short sentences.
+- Embedding: Generates embeddings for two short sentences.
 
 ## Usage
 
@@ -50,7 +50,7 @@ Note: `.env.local` takes precedence over `.env` if both exist.
 - `GEMINI_API_KEY` or `GOOGLE_API_KEY` - Required for Gemini tests
 - `OLLAMA_MODEL` - Optional, defaults to "smollm" for Ollama tests
 - `OLLAMA_SUMMARY_MODEL` - Optional override for summarize tests (defaults to `OLLAMA_MODEL`)
-- `OLLAMA_EMBED_MODEL` - Optional override for embed tests (defaults to `nomic-embed-text`)
+- `OLLAMA_EMBEDDING_MODEL` - Optional override for embedding tests (defaults to `nomic-embed-text`)
 
 Tests will be skipped for adapters where the API key is not set.
 
@@ -62,7 +62,7 @@ Each test run creates detailed debug files in the `output/` directory:
 - `{adapter}-test2-temperature-tool.json` - Debug info for tool call
 - `{adapter}-test3-approval-tool-flow.json` - Debug info for approval flow
 - `{adapter}-test5-summarize.json` - Debug info for summarize
-- `{adapter}-test6-embed.json` - Debug info for embed
+- `{adapter}-test6-embedding.json` - Debug info for embedding
 
 Each debug file contains:
