+++ packages/typescript/tests-adapters/env.example
@@ -6,10 +6,10 @@ GEMINI_API_KEY=your_gemini_api_key_here
 # Ollama models (optional, defaults shown)
 OLLAMA_MODEL=smollm
 OLLAMA_SUMMARY_MODEL=smollm
-OLLAMA_EMBED_MODEL=nomic-embed-text
+OLLAMA_EMBEDDING_MODEL=nomic-embed-text
 
 # Optional overrides for embeddings/summarization on hosted providers
-OPENAI_EMBED_MODEL=text-embedding-3-small
+OPENAI_EMBEDDING_MODEL=text-embedding-3-small
 OPENAI_SUMMARY_MODEL=gpt-4o-mini
-GEMINI_EMBED_MODEL=embedding-001
+GEMINI_EMBEDDING_MODEL=embedding-001
 GEMINI_SUMMARY_MODEL=gemini-2.5-flash
